OnTrigger Call Change
Granted at least one collider is a trigger the OnTrigger call back is execute for both colliders.

If a player (Non-trigger) walks into a coin (Trigger), the coin's ontrigger will be called taking in the player collider as the "other" and the players ontrigger will be called taking in the coins collider as the "other".