Bunch of Random Changes
Better safety for AssetHandler getters (Won't automatically create an empty entry if the asset the person is trying to get is not on the map.

Automatic aspect ratio update based on window resizing.

EMPTY static const for Vertex and Constant Data.

Re-named misleading skyboxTextureName to reflectedTextureName.

Made sure cameras roll is properly initialized so the world isn't upside down.